remove warnings on windows

+++ mountain/public_inc/mountain/sampler.h
@@ -13,7 +13,7 @@ namespace mountain::image {
          * @param context: Vulkan context
          * @param mipmap_levels: mipmap level you want
          */
-        MOUNTAINAPI_EXPORT sampler(Context const &context, float mipmap_levels);
+        MOUNTAINAPI_EXPORT sampler(Context const &context, uint32_t mipmap_levels);
 
 #ifndef DOXYGEN_SHOULD_SKIP_THIS
         explicit operator vk::Sampler() const{return *_sampler;}
